ived made some small edits to the models for photos to amke correct validations

+++ photos/models.py
@@ -10,7 +10,7 @@ class Photo(models.Model):
 
     Location = models.CharField(max_length=100)
     desc = models.TextField()
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
     has_sizes = models.BooleanField(default=False, null=True, blank=True)
     img = models.ImageField(upload_to='media')
     NE = models.BooleanField(default=True)
